pyproject.toml
[tool.poetry]
name = "belx"
version = "0.1.0"
description = ""
authors = ["tsutomu-nagano <copyxshow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
line-bot-sdk = "^2.1.0"
beautifulsoup4 = "^4.10.0"
redis = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
